requirements.txt
httpx>=0.23.0
supabase>=2.4.0
mcp>=1.0.0
pydantic>=2.0.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0